+++ RELEASE_NOTES.md
@@ -6,15 +6,15 @@
 
 ## WORK IN PROGRESS
 
-> List updated Jun 21 2022
+> List updated Jun 29 2022
 
 **⚠ PLEASE UNDERSTAND:** Features and changes below may not fully work as the source code is constantly under development. For a stable and more tested version, please use published binary releases below.
 
 **Completed items**:
-* Minor fixes to registration view (remove Discord, not needed) and enable email checkbox always
+
 
 **Waiting**:
-* Rundown Variables (does not remove prVar from profile if no users)
+- Rundown Variables (does not remove prVar from profile if no users)
         
 <BR>
 
@@ -27,6 +27,7 @@ Published releases:
 ## **1.1.3** (Jun 29 2022)
 > Version 1.1.3 was never released publicly as binaries. These features will be present in the future releases.
 - New url parameter added to the renderer to pass a desired refresh rate to templates `renderer/?fps=30`. This will require modifications to the templates with controllable update rates, such as the scrolling tickers, using **gsap** animation library (or similar.) See _Google Sheet Ticker_ -template for an example.
+- Minor fixes to registration view (remove Discord, not needed) and enable email checkbox always
 
 ## **1.1.2** (Jun 20 2022)
 * >See also [Latest changes Knowledge Base article](https://spxgc.tawk.help/article/latest-changes) for selected feature highlights.